requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pyjwt==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.6
requests==2.32.3
bcrypt==4.2.1
openai==1.58.1
